Fix errors with mason-lspconfig by removing setup function for mason-lspconfig
Upon studying the documentation for mason-lspconfig very carefully,
I realized that the setup() function for that plugin didn't need to be included.
See this link:
https://github.com/mason-org/mason-lspconfig.nvim/tree/main?tab=readme-ov-file#recommended-setup-for-lazynvim

"The following is the recommended setup when using lazy.nvim. It will set up the plugin for you, meaning you don't have to call require("mason-lspconfig").setup() yourself."

After removing the setup for mason-lspconfig, everything seems to work!
I also replaced williamboman with mason-org, which seems to be the most up-to-date address.